+++ Healthcare/src/main/java/com/cafe24/kyungsu93/medication/controller/MedicationController.java
@@ -14,6 +14,8 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.cafe24.kyungsu93.medication.service.MedicationService;
+import com.cafe24.kyungsu93.member.service.Member;
+import com.cafe24.kyungsu93.medication.service.MedicationDao;
 import com.cafe24.kyungsu93.medication.service.MedicationRequest;
 
 @Controller
@@ -27,55 +29,78 @@ public class MedicationController {
 									,HttpSession session
 									,MedicationRequest medicationRequest
 									,@RequestParam(value="currentPage", defaultValue="1" ) int currentPage
-									,@RequestParam(value="pagePerRow", defaultValue="10" ) int pagePerRow) {		
+									,@RequestParam(value="pagePerRow", defaultValue="10" ) int pagePerRow) {	
 		logger.debug("MedicationController.getMedicationList GET 방식 호출");
-		medicationRequest.setMemberNo(session.getAttribute("memberSessionNo").toString());
-		Map map = medicationService.getMedicationList(currentPage, pagePerRow, medicationRequest);
-		model.addAttribute("list", map.get("list"));
-		model.addAttribute("firstPage", map.get("firstPage"));
-		model.addAttribute("lastPage", map.get("lastPage"));
-		model.addAttribute("beforePage", map.get("beforePage"));
-		model.addAttribute("afterPage", map.get("afterPage"));
-		model.addAttribute("currentPage", currentPage);
-		model.addAttribute("pagePerRow", pagePerRow);
-		return "medication/getMedicationList";
+		if(session.getAttribute("memberSessionLevel") != null && session.getAttribute("memberSessionLevel").toString().equals("2")) {
+			medicationRequest.setMemberNo(session.getAttribute("memberSessionNo").toString());
+			Map map = medicationService.getMedicationList(currentPage, pagePerRow, medicationRequest);
+			model.addAttribute("list", map.get("list"));
+			model.addAttribute("firstPage", map.get("firstPage"));
+			model.addAttribute("lastPage", map.get("lastPage"));
+			model.addAttribute("beforePage", map.get("beforePage"));
+			model.addAttribute("afterPage", map.get("afterPage"));
+			model.addAttribute("currentPage", currentPage);
+			model.addAttribute("pagePerRow", pagePerRow);
+			return "medication/getMedicationList";
+		} else {
+			return "redirect:/";
+		}
+
 	}
 	
 	@RequestMapping(value="/addMedication", method=RequestMethod.GET)
 	public String addMedication(Model model
 			,@RequestParam(value="currentPage", defaultValue="1" ) int currentPage
-			,@RequestParam(value="pagePerRow", defaultValue="10" ) int pagePerRow) {
+			,@RequestParam(value="pagePerRow", defaultValue="10" ) int pagePerRow
+			,HttpSession session) {
 		logger.debug("MedicationController.addMedication GET 방식 호출");
-		model.addAttribute("currentPage", currentPage);
-		model.addAttribute("pagePerRow", pagePerRow);
-		return "medication/addMedication";
+		if(session.getAttribute("memberSessionLevel").toString().equals("2")) {
+			model.addAttribute("currentPage", currentPage);
+			model.addAttribute("pagePerRow", pagePerRow);
+			return "medication/addMedication";
+		} else {
+			return "redirect:/";
+		}
+		
 	}
 	
 	@RequestMapping(value="/addMedication", method=RequestMethod.POST)
 	public String addMedication(MedicationRequest medicationRequest, HttpSession session, Model model) {
 		logger.debug("MedicationController.addMedication POST 방식 호출");
-		medicationRequest.setMemberNo(session.getAttribute("memberSessionNo").toString());
-		medicationService.addMedication(medicationRequest);
-		return "redirect:/getMedicationList";
+		if(session.getAttribute("memberSessionLevel").toString().equals("2")) {
+			medicationRequest.setMemberNo(session.getAttribute("memberSessionNo").toString());
+			medicationService.addMedication(medicationRequest);
+			return "redirect:/getMedicationList";
+		} else {
+			return "redirect:/";
+		}
 	}
 	
 	@RequestMapping(value="/getMedicationContent", method=RequestMethod.GET)
 	public String getMedicationContent(MedicationRequest medicationRequest, Model model
-									,@RequestParam(value="currentPage") int currentPage
-									,@RequestParam(value="pagePerRow", defaultValue="10" ) int pagePerRow) {
+										,@RequestParam(value="currentPage") int currentPage
+										,@RequestParam(value="pagePerRow", defaultValue="10" ) int pagePerRow
+										,HttpSession session) {
 		logger.debug("MedicationController.addMedication POST 방식 호출");
-		model.addAttribute("currentPage", currentPage);
-		model.addAttribute("pagePerRow", pagePerRow);
-		model.addAttribute("medicationResponse", medicationService.getMedicationContent(medicationRequest));
-		return "medication/getMedicationContent";
+		if(session.getAttribute("memberSessionNo").toString().equals(medicationService.getMedicationContent(medicationRequest).getMemberNo())) {
+			model.addAttribute("currentPage", currentPage);
+			model.addAttribute("pagePerRow", pagePerRow);
+			model.addAttribute("medicationResponse", medicationService.getMedicationContent(medicationRequest));
+			return "medication/getMedicationContent";
+		} else {
+			return "redirect:/";
+		}
 	}
 	
 	@RequestMapping(value="/removeMedication", method=RequestMethod.GET)
 	public String removeMedication(MedicationRequest medicationRequest
 									,@RequestParam(value="currentPage") int currentPage
-									,@RequestParam(value="pagePerRow", defaultValue="10" ) int pagePerRow) {
+									,@RequestParam(value="pagePerRow", defaultValue="10" ) int pagePerRow
+									,HttpSession session) {
 		logger.debug("MedicationController.removeMedication GET 방식 호출");
-		medicationService.removeMedication(medicationRequest);
+		if(session.getAttribute("memberSessionNo").toString().equals(medicationService.getMedicationContent(medicationRequest).getMemberNo())) {
+			medicationService.removeMedication(medicationRequest);
+		}
 		return "redirect:/getMedicationList";
 	}
 	
@@ -85,12 +110,17 @@ public class MedicationController {
 	public String modifyMedication(Model model
 									,MedicationRequest medicationRequest
 									,@RequestParam(value="currentPage") int currentPage
-									,@RequestParam(value="pagePerRow", defaultValue="10" ) int pagePerRow) {
+									,@RequestParam(value="pagePerRow", defaultValue="10" ) int pagePerRow
+									,HttpSession session) {
 		logger.debug("MedicationController.modifyMedication GET 방식 호출");
-		model.addAttribute("medicationResponse",medicationService.getMedicationContent(medicationRequest));
-		model.addAttribute("currentPage", currentPage);
-		model.addAttribute("pagePerRow", pagePerRow);
-		return "medication/modifyMedication";
+		if(session.getAttribute("memberSessionNo").toString().equals(medicationService.getMedicationContent(medicationRequest).getMemberNo())) {
+			model.addAttribute("medicationResponse",medicationService.getMedicationContent(medicationRequest));
+			model.addAttribute("currentPage", currentPage);
+			model.addAttribute("pagePerRow", pagePerRow);
+			return "medication/modifyMedication";
+		} else {
+			return "redirect:/getMedicationList";
+		}
 	}
 	
 	@RequestMapping(value="/modifyMedication", method=RequestMethod.POST)
@@ -100,7 +130,11 @@ public class MedicationController {
 									,@RequestParam(value="currentPage") int currentPage
 									,@RequestParam(value="pagePerRow", defaultValue="10" ) int pagePerRow) {
 		logger.debug("MedicationController.modifyMedication POST 방식 호출");
-		medicationService.modifyMedication(medicationRequest);
-		return "redirect:/getMedicationContent?medicationNo=" + medicationRequest.getMedicationNo() + "&currentPage=" + currentPage + "&pagePerRow=" + pagePerRow;
+		if(session.getAttribute("memberSessionNo").toString().equals(medicationService.getMedicationContent(medicationRequest).getMemberNo())) {
+			medicationService.modifyMedication(medicationRequest);
+			return "redirect:/getMedicationContent?medicationNo=" + medicationRequest.getMedicationNo() + "&currentPage=" + currentPage + "&pagePerRow=" + pagePerRow;
+		} else {
+			return "redirect:/getMedicationList";
+		}
 	}
 }
